Refactor UiText location
