+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/data/repositories/syncRepository/syncRepositoryImpls/SyncRepositoryImpl.kt
@@ -10,7 +10,7 @@ import com.realityexpander.tasky.agenda_feature.data.repositories.syncRepository
 import com.realityexpander.tasky.agenda_feature.data.repositories.syncRepository.remote.SyncAgendaRequestDTO
 import com.realityexpander.tasky.agenda_feature.domain.AgendaItem
 import com.realityexpander.tasky.agenda_feature.domain.ResultUiText
-import com.realityexpander.tasky.core.presentation.common.util.UiText
+import com.realityexpander.tasky.core.presentation.util.UiText
 import javax.inject.Inject
 
 class SyncRepositoryImpl @Inject constructor(
