+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/presentation/reminder_screen/ReminderViewModel.kt
@@ -12,7 +12,7 @@ import com.realityexpander.tasky.agenda_feature.presentation.reminder_screen.Rem
 import com.realityexpander.tasky.auth_feature.domain.IAuthRepository
 import com.realityexpander.tasky.auth_feature.domain.validation.ValidateEmail
 import com.realityexpander.tasky.core.presentation.common.SavedStateConstants
-import com.realityexpander.tasky.core.presentation.common.util.UiText
+import com.realityexpander.tasky.core.presentation.util.UiText
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.launch
