+++ app/src/main/java/com/realityexpander/tasky/auth_feature/presentation/components/TextEntryField.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import com.realityexpander.tasky.R
-import com.realityexpander.tasky.core.presentation.common.util.UiText
+import com.realityexpander.tasky.core.presentation.util.UiText
 import com.realityexpander.tasky.core.presentation.theme.textEntryFieldTextStyle
 
 @Composable
