+++ app/src/main/res/values/strings.xml
@@ -217,5 +217,6 @@
     <string name="reminder_error_delete_reminder">Error deleting Reminder</string>
     <string name="reminder_message_reminder_saved">Reminder Saved</string>
     <string name="reminder_message_reminder_deleted_success">Reminder successfully deleted</string>
+    <string name="no_internet">Network Offline</string>
 
 </resources>
